+++ Sources/AudioIO/PitchShifterType.swift
@@ -1,0 +1,12 @@
+//
+//  PitchShifterType.swift
+//  AudioIO
+//
+//  Created by Sam Meech-Ward on 2018-01-16.
+//
+
+import Foundation
+
+public protocol PitchShifterType: class {
+  var pitch: Double { get set }
+}